requirements.txt
github3.py==1.1.0
requests==2.19.1